♻️ refactor(Home): Home 화면에 들어가는 Main 컴포넌트화
Issue #6
