+++ client/src/components/homepage/index.js
@@ -2,3 +2,4 @@ export * from './Dot';
 export * from './MemberInfo';
 export * from './Header';
 export * from './EnterButton';
+export * from './Main';
